refactoring and fixing breaking changes.
