+++ client/helpers/getStringElipsis.js
