+++ client/helpers/mapDomainDescription.js
