+++ client/routes/execution/constants.js
@@ -1,6 +1,1 @@
-export const jsonKeys = ['result', 'input', 'details', 'data', 'Error'];
-export const preKeys = jsonKeys.concat(['stackTrace', 'details.stackTrace']);
-
-export const MAXIMUM_JSON_CHARACTER_LIMIT = 5000;
-export const MAXIMUM_JSON_MESSAGE = '\n ... to see more open full screen mode from top right arrow.';
 export const RESULT_THRESHOLD = 1000;
