+++ client/routes/execution/helpers/getEventSummary.js
@@ -1,6 +1,6 @@
 // TODO - move these transforms into helpers folder
 import { summarizeEvents } from './summarizeEvents';
-import getKeyValuePairs from './getKeyValuePairs';
+import { getKeyValuePairs } from '../../../helpers';
 
 const getEventSummary = event => {
   if (!event) {
